Rename class to BiomeConfigReaderV1, interface to BiomeConfigReader, consolidate getClosestBiomeConfigPath
